requirements.txt
psycopg2-binary==2.9.9
python-dotenv==1.0.0
requests==2.31.0
anthropic==0.43.1
apify-client==2.2.1
httpx==0.27.2
zhipuai==2.1.5.20250825
